Indian_Finance_Agent/requirements.txt
crewai
langchain_google_genai
langchain_community
duckduckgo-search
yfinance
python-dotenv
langchain_openai
fastapi
uvicorn
streamlit
arize-phoenix
openinference-instrumentation-crewai
